major: assets should not be parsed (#304)
This PR changes the behavior of dependency tracing such that `fs.readFile('./file.js')` is no longer considered a `dependency` and it will not be traced. Instead, it is considered an `asset` the same way a `fs.readFile('./file.txt')` is. The `dependency` type will be reserved for `import` and `require` only.

- Fixes #303 
- Related to https://github.com/vercel/customer-issues/issues/608

Note that this could be considered a breaking change since https://github.com/vercel/nft/pull/40 expected the `readFile()`  config to be called for assets but it should only be called for dependencies.

In order to reduce the amount of breakage, `require.resolve()` was left as dependency instead of asset since its often passed to `child_process.spawn()` which needs the dependencies.

Co-authored-by: JJ Kasper <jj@jjsweb.site>