fix: Make caching work correctly in async context
BREAKING CHANGE: The file-IO caches now contains promises instead of the actual result.

Previously the cache was only populated after the function call succeed which in a async
context was leading to multiple file IO or analyze tasks started for the same file
at once. Caching the Promise instead of the result makes this problem go away.