fix: Make caching work correctly in async context (#320)
Right now the cache was only populated after the call to the fs API
succeed, which in an async context was leading to multiple file IO or
analyze tasks started for the same file at once. This is because both
file IO are async and so until the fs call finished other calls for the
same file can still come in.

To fix this I extracted all file IO calls into a separate class that
wraps everything correctly and instead of caching the actual result, now
a Promise is cached. So the cache now always needs to be awaited.

As the values in the cache change this is definitely a breaking change.
Something like this:

```diff
const cache = {};
await nodeFileTrace([file], {
    cache,
);

-const cacheValue = cache.fileCache.get(file);
+const cacheValue = await cache.fileCache.get(file);
```

I know this is quite a big change, just let me know what you think and
if I should change anything. :)

In my personal example the number of calls to analyze went from ~2400 to
~1600 and the runtime from ~3s to ~2.7s (which includes other work than
nft)

Co-authored-by: Steven <steven@ceriously.com>