feat(resolve): export `resolve()` function (#354)
nft allows you to override the resolve step with a hook however it does
not expose its internal resolveDependency implementation. This means if
you want to take over resolving in certain cases you must do so in every
case. To make conditional or augmented resolving possible this change
reexports the resolveDependency function from the index. This way you
can use it in custom resolve hook implementations

---------

Co-authored-by: Steven <steven@ceriously.com>