feat: support import attributes (#377)
The [Import Attributes
proposal](https://github.com/tc39/proposal-import-attributes) is in
Stage 3, so implementation is started. Currently, NFT bails on any file
that includes import attributes, because Acorn can't parse them. That's
because Acorn only supports Stage 4 proposals - but for import
attributes, there's a plugin that this PR adds to NFT.