fix: Handle the scenario where a package.json#browser field could be `false` (#427)
We [recently shipped](https://github.com/vercel/nft/pull/424) support
for the `browser` field, but we discovered an edge case where instead of
mapping to another string, the value could be `false` (eg,
[`object-inspect`](https://github.com/inspect-js/object-inspect/blob/main/package.json#L82)),
which indicates that this module should be treated as an empty object
(`{}`) (evidenced by its usage
[here](https://github.com/inspect-js/object-inspect/blob/main/index.js#L68-L70)).

The code was previously erroring on the string's `startsWith` checks,
causing the entire package to be omitted from the file mapping result.
This ensures that if the value is `false` we'll continue on.

---------

Co-authored-by: Steven <steven@ceriously.com>
Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>